scripts/requirements.txt
beautifulsoup4==4.10.0
black==21.12b0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
Deprecated==1.2.13
idna==3.3
lxml==4.7.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
pycparser==2.21
PyGithub==1.55
PyJWT==2.3.0
PyNaCl==1.5.0
python-dotenv==0.19.2
requests==2.27.1
soupsieve==2.3.1
tomli==1.2.3
typing_extensions==4.0.1
urllib3==1.26.8
wrapt==1.13.3
